Подготовка к деплою

+++ bot/core/models/guild_settings.py
@@ -1,3 +1,4 @@
+from sqlalchemy import BIGINT
 from sqlalchemy.orm import mapped_column, Mapped
 
 from .base import Base
@@ -5,11 +6,11 @@ from .base import Base
 
 class GuildSettings(Base):
     __tablename__ = "guilds_settings"
-    id: Mapped[int] = mapped_column(primary_key=True, index=True)
-    guild_id: Mapped[int] = mapped_column(nullable=False, unique=True)
-    greetings_channel_id: Mapped[int] = mapped_column(nullable=True)
-    ai_channel_id: Mapped[int] = mapped_column(nullable=True)
-    activist_role_id: Mapped[int] = mapped_column(nullable=True)
+    id: Mapped[int] = mapped_column(BIGINT(), primary_key=True, index=True)
+    guild_id: Mapped[int] = mapped_column(BIGINT(), nullable=False, unique=True)
+    greetings_channel_id: Mapped[int] = mapped_column(BIGINT(), nullable=True)
+    ai_channel_id: Mapped[int] = mapped_column(BIGINT(), nullable=True)
+    activist_role_id: Mapped[int] = mapped_column(BIGINT(), nullable=True)
     activist_role_messages_count: Mapped[int] = mapped_column(nullable=True)
     """Сколько сообщений надо отправить чтобы получить роль активиста"""
     is_greetings_enabled: Mapped[bool] = mapped_column(default=False)
